futurevox/requirements.txt
torch>=1.12.0
torchaudio>=0.12.0
pytorch-lightning>=1.7.0
matplotlib>=3.5.0
numpy>=1.20.0
pyyaml>=6.0
omegaconf>=2.1.1
librosa>=0.9.2
phonemizer>=3.2.0
inflect>=5.6.0
tqdm>=4.62.0
einops>=0.4.1
scipy>=1.7.0
tensorboard>=2.8.0
soundfile>=0.10.3
pillow>=9.0.0
pyworld>=0.3.0